Correct postgres bulk-stream loading
Fix #20

The thread copying the data could have ended up in error before
the end of the copied data, making the data-generation thread
push data into the stream without it being consumed.
Now the main thread checks for the copying thread to be alive every
thousand row, and stops in case of error.